requirements.txt
discord.py[voice]>=2.4.0
yt-dlp>=2024.12.0
python-dotenv>=1.0.0
